Add support for multiple generators
