+++ bbx_sandbox/examples/02_sine_wave_overdrive.rs
@@ -4,7 +4,10 @@ use bbx_sandbox::{constants::SAMPLE_RATE, player::Player, signal::Signal};
 pub fn create_graph() -> Graph {
     let mut graph = Graph::new(SAMPLE_RATE);
 
-    let oscillator = graph.add_generator(Generator::new(SAMPLE_RATE, Some(110.0)));
+    let oscillator = graph.add_generator(Generator::Wavetable {
+        sample_rate: SAMPLE_RATE,
+        frequency: 110.0,
+    });
     let overdrive = graph.add_effector(Effector::Overdrive());
     graph.create_connection(oscillator, overdrive);
